requirements.txt
flask==1.1.1
flask-wtf==0.14.3
Flask-sqlalchemy==0.8.0